refactor(tasks): give the sub-task list its own component
TaskRowDescription had grown two jobs: the sub-task fetch with its four states
(loading, rows, empty-or-failed, load-more) and the linked page's document. The
sub-task half moves to TaskSubTaskList unchanged — same markup, same states,
same tests passing — leaving TaskRowDescription to compose the two halves.

The seam is the one the epic needs next: task 5 branches between showing
sub-tasks and showing the document, which is now a choice between two
components rather than surgery inside one.

The render tests still drive TaskRowDescription rather than the new component
directly. What matters is what the assembled expansion puts on screen, and going
through the composition means they survive the two halves being rearranged.
